backend/requirements.txt
# API Framework
fastapi==0.111.0
uvicorn==0.30.1

# OpenAI for LLM
openai==1.44.0

# Embedding and Semantic Search
faiss-cpu==1.7.4
sentence-transformers==2.7.0

# File Handling (PDF, Word)
PyMuPDF==1.24.1
python-docx==1.0.1

# Utilities
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4

# Validation & Upload
pydantic==2.7.0
python-multipart==0.0.9

# Testing
pytest==8.2.1

# Environment variables
python-dotenv==1.0.1
#faiss-cpu
faiss-cpu==1.8.0
